+++ CHANGELOG.md
@@ -3,13 +3,15 @@
 ## Unreleased
 ### ADDED
 
-- Add an amino acid nucleotide quilt plot by [@mbowcut2](https://github.com/pinellolab/CRISPResso2/pull/552) in [#552](https://github.com/pinellolab/CRISPResso2/pull/552)
+- Add an amino acid nucleotide quilt plot by [@mbowcut2](https://github.com/mbowcut2) in [#552](https://github.com/pinellolab/CRISPResso2/pull/552)
 
 - Add `scripts/reconstituteReads.py` to generate FASTQ from CRISPResso2 output by [@kclem](https://github.com/kclem) in [`a800762`](https://github.com/pinellolab/CRISPResso2/commit/a800762712e692a9fc7005b3510d013924c843a5) and [`cd79dcc`](https://github.com/pinellolab/CRISPResso2/commit/cd79dcc3ca9c06736dc6cd409d43dee8e4d1ae69)
 
 - Add an [UpSet plot](https://upset.app/) to represent bystander edits for Base Editing analyses by [@mbowcut2](https://github.com/mbowcut2) in [#554](https://github.com/pinellolab/CRISPResso2/pull/554)
 
 - Allow for messages to be served via CRISPResso reports by [@Colelyman](https://github.com/Colelyman) in [#583](https://github.com/pinellolab/CRISPResso2/pull/583)
+
+- Add a plot that shows the distribution of homology scores for reads by [@mbowcut2](https://github.com/mbowcut2) in [#601](https://github.com/pinellolab/CRISPResso2/pull/601)
 
 ### FIXED